Create files and directories

s3_file_create create file on AWS S3, if file already exists it will be left unchanged.

s3_dir_create create empty directory of AWS S3.

Usage

s3_file_create(path, overwrite = FALSE, ...)

s3_bucket_create(
  path,
  region_name = NULL,
  mode = c("private", "public-read", "public-read-write", "authenticated-read"),
  versioning = FALSE,
  ...
)

s3_dir_create(path, overwrite = FALSE, ...)

Arguments

path

(character): A character vector of path or s3 uri.

overwrite

(logical): Overwrite files if the exist. If this is FALSE and the file exists an error will be thrown.

...

parameters to be passed to s3_put_object, s3_create_bucket

region_name

(character): region for AWS S3 bucket, defaults to s3_file_system() class region.

mode

(character): A character of the mode

versioning

(logical)

Value

character vector of s3 uri paths

Examples

if (FALSE) { # \dontrun{
# Require AWS S3 credentials

temp_file = s3_file_temp(tmp_dir= "MyBucket")
s3_file_create(temp_file)
} # }