Get files and directories information

s3_file_info returns file information within AWS S3 directory

s3_file_size returns file size in bytes

s3_dir_info returns file name information within AWS S3 directory

s3_dir_ls returns file name within AWS S3 directory

s3_file_info(path)

s3_file_size(path)

s3_dir_info(
  path = ".",
  type = c("any", "bucket", "directory", "file"),
  glob = NULL,
  regexp = NULL,
  invert = FALSE,
  recurse = FALSE,
  refresh = FALSE,
  ...
)

s3_dir_ls(
  path = ".",
  type = c("any", "bucket", "directory", "file"),
  glob = NULL,
  regexp = NULL,
  invert = FALSE,
  recurse = FALSE,
  refresh = FALSE,
  ...
)

Arguments

path

(character):A character vector of one or more paths. Can be path or s3 uri.

type

(character): File type(s) to return. Default ("any") returns all AWS S3 object types.

glob

(character): A wildcard pattern (e.g. *.csv), passed onto grep() to filter paths.

regexp

(character): A regular expression (e.g. [.]csv$), passed onto grep() to filter paths.

invert

(logical): If code return files which do not match.

recurse

(logical): Returns all AWS S3 objects in lower sub directories

refresh

(logical): Refresh cached in s3_cache.

...

parameters to be passed to s3_list_objects_v2

Value

s3_file_info A data.table with metadata for each file. Columns returned are as follows.

bucket_name (character): AWS S3 bucket of file

key (character): AWS S3 path key of file

uri (character): S3 uri of file

size (numeric): file size in bytes

type (character): file type (file or directory)

etag (character): An entity tag is an opague identifier

last_modified (POSIXct): Created date of file.

delete_marker (logical): Specifies retrieved a logical marker

accept_ranges (character): Indicates that a range of bytes was specified.

expiration (character): File expiration

restore (character): If file is archived

archive_status (character): Archive status

missing_meta (integer): Number of metadata entries not returned in "x-amz-meta" headers

version_id (character): version id of file

cache_control (character): caching behaviour for the request/reply chain

content_disposition (character): presentational information of file

content_encoding (character): file content encodings

content_language (character): what language the content is in

content_type (character): file MIME type

expires (POSIXct): date and time the file is no longer cacheable

website_redirect_location (character): redirecs request for file to another

server_side_encryption (character): File server side encryption

metadata (list): metadata of file

sse_customer_algorithm (character): server-side encryption with a customer-provided encryption key

sse_customer_key_md5 (character): server-side encryption with a customer-provided encryption key

ssekms_key_id (character): ID of the Amazon Web Services Key Management Service

bucket_key_enabled (logical): s3 bucket key for server-side encryption with

storage_class (character): file storage class information

request_charged (character): indicates successfully charged for request

replication_status (character): return specific header if request involves a bucket that is either a source or a destination in a replication rule https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.head_object

parts_count (integer): number of count parts the file has

object_lock_mode (character): the file lock mode

object_lock_retain_until_date (POSIXct): date and time of when object_lock_mode expires

object_lock_legal_hold_status (character): file legal holding

s3_dir_info data.table with directory metadata

bucket_name (character): AWS S3 bucket of file

key (character): AWS S3 path key of file

uri (character): S3 uri of file

size (numeric): file size in bytes

version_id (character): version id of file

etag (character): An entity tag is an opague identifier

last_modified (POSIXct): Created date of file

s3_dir_ls character vector of s3 uri paths

Examples

if (FALSE) {
# Require AWS S3 credentials

temp_file = s3_file_temp(tmp_dir= "MyBucket")
s3_file_create(temp_file)

s3_file_info(temp_file)
}